// src/features/order/models/order.views/requestView.model.ts
export class RequestViewModel {
  REQ_RAZN_KEY: number;
  VID_TEHN: string;
  DATE_RAB: Date;
  VR_OT: string;
  VR_DO: string;
  KOL: number;
  MESTO_POD: string;
  TRASSA: string;
  OTV: string | null;
  ZAK: string | null;
  COMMENT: string | null;
  PRIM_DISP: string;
  STAMP: string;
  RES: number;
  REQ_USER_ID: number;
  LNAME: string;
  PODR: string;
  TRASSA_R: string;
  ZAK_ID: number;
  PODR_ID: number;
  ADR_POGR: string;
  ADR_RAZGR: string;
  TIP_GRUZ: string;
  VES: number | null;
  RAST_KM: number | null;
  HODOK: number | null;
  RAZN_ID: number;
  NOM_REQ: string;
  FIO_STROP: string;
  NOM_DOC_STROP: string;
  DATE_DOC_STROP: string | null;
  FIO_OTV: string | null;
  NOM_DOC_OTV: string | null;
  DATE_DOC_OTV: string | null;
  PRIOR: number | null;
  PRIOR_STR: string | null;
  XID: string | null;
  DISLOCATION_GUID: string | null;
}